Добавлено: аутентификация и разделение ролей пользователей, хранение идентификатора пользователя в точке входа, просмотр и извенение текущего профиля, завершение текущего сеанса.

+++ src/main/java/ru/evasmall/tm/service/ProjectService.java
@@ -1,5 +1,6 @@
 package ru.evasmall.tm.service;
 
+import ru.evasmall.tm.Application;
 import ru.evasmall.tm.entity.Project;
 import ru.evasmall.tm.repository.ProjectRepository;
 
@@ -8,6 +9,8 @@ import java.util.List;
 public class ProjectService {
 
     private final ProjectRepository projectRepository;
+
+
 
     public ProjectService(ProjectRepository projectRepository) {
         this.projectRepository = projectRepository;
@@ -22,10 +25,10 @@ public class ProjectService {
         return projectRepository.create(name);
     }
 
-    public Project create(String name, String description) {
+    public Project create(String name, String description, Long userid) {
         if (name == null || name.isEmpty()) return null;
         if (description == null || description.isEmpty()) return null;
-        return projectRepository.create(name, description);
+        return projectRepository.create(name, description, userid);
     }
 
     public Project update(Long id, String name, String description) {
@@ -39,34 +42,90 @@ public class ProjectService {
         projectRepository.clear();
     }
 
+    //Поиск проекта по индексу
     public Project findByIndex(int index) {
         if (index < 0 || index > projectRepository.size() - 1) return null;
         return projectRepository.findByIndex(index);
     }
 
+    //Поиск проекта по индексу с учетом принадлежности пользователю текущей сессии
+    public Project findByIndexUserId(int index) {
+        if (index < 0 || index > projectRepository.size() - 1) return null;
+        if (projectRepository.findByIndex(index) == null) return null;
+        if (projectRepository.findByIndex(index).getUserid().equals(Application.userIdCurrent)) return projectRepository.findByIndex(index);
+        else return null;
+    }
+
+    //Поиск проекта по наименованию
     public Project findByName(String name) {
         if (name == null || name.isEmpty()) return null;
         return projectRepository.findByName(name);
     }
 
+    //Поиск проекта по наименованию с учетом принадлежности пользователю текущей сессии
+    public Project findByNameUserId(String name) {
+        if (name == null || name.isEmpty()) return null;
+        if (projectRepository.findByName(name) == null) return null;
+        if (projectRepository.findByName(name).getUserid().equals(Application.userIdCurrent)) return projectRepository.findByName(name);
+        else return null;
+    }
+
+    //Поиск проекта по идентификатору
     public Project findById(Long id) {
         if (id == 0) return null;
+        if (projectRepository.findById(id) == null) return  null;
         return projectRepository.findById(id);
     }
 
+    //Поиск проекта по идентификатору с учетом принадлежности пользователю текущей сессии
+    public Project findByIdUserId(Long id) {
+        if (id == 0) return null;
+        if (projectRepository.findById(id) == null) return  null;
+        if (projectRepository.findById(id).getUserid().equals(Application.userIdCurrent)) return projectRepository.findById(id);
+        return null;
+    }
+
+    //Удаление проекта по наименованию
     public Project removeByName(String name) {
         if (name == null || name.isEmpty()) return null;
         return projectRepository.removeByName(name);
     }
 
+    //Удаление проекта по наименованию с учетом принадлежности пользователю текущей сессии
+    public Project removeByNameUserId(String name) {
+        if (name == null || name.isEmpty()) return null;
+        if (projectRepository.findByName(name) == null) return null;
+        if (projectRepository.findByName(name).getUserid().equals(Application.userIdCurrent)) return projectRepository.removeByName(name);
+        else return null;
+    }
+
+    //Удаление проекта по идентификатору
     public Project removeById(Long id) {
         if (id == null ) return null;
+        if (projectRepository.findById(id) == null) return  null;
         return projectRepository.removeById(id);
     }
 
+    //Удаление проекта по идентификатору с учетом принадлежности пользователю текущей сессии
+    public Project removeByIdUserId(Long id) {
+        if (id == null ) return null;
+        if (projectRepository.findById(id) == null) return  null;
+        if (projectRepository.findById(id).getUserid().equals(Application.userIdCurrent)) return projectRepository.removeById(id);
+        return null;
+    }
+
+    //Удаление проекта по индексу
     public Project removeByIndex(int index) {
         if (index < 0 || index > projectRepository.size() -1) return null;
         return projectRepository.removeByIndex(index);
+    }
+
+    //Удаление проекта по индексу с учетом принадлежности пользователю текущей сессии
+    public Project removeByIndexUserId(int index) {
+        if (index < 0 || index > projectRepository.size() -1) return null;
+        if (projectRepository.findByIndex(index) == null) return null;
+        if (projectRepository.findByIndex(index).getUserid().equals(Application.userIdCurrent)) return projectRepository.removeByIndex(index);
+        else return null;
     }
 
 }
